+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.9.0]
+- [Tracing working, Correct tracestate spelling, pass on empty tracestate if none is provided #389](https://github.com/xmidt-org/tr1d1um/pull/389)
+
+## [v0.8.5]
+- [Add Trace info to WRP Message Header #385](https://github.com/xmidt-org/tr1d1um/pull/385)
+- [Update Tracing Configurations #386](https://github.com/xmidt-org/tr1d1um/pull/386)
+
+## [v0.8.4]
+- Update dependencies
+
 ## [v0.8.3]
 - [Bug: Event related metrics fail to update #361](https://github.com/xmidt-org/tr1d1um/issues/361)
 - [Tr1d1um Panic: device status code used as the response status code can cause an unrecoverable panic #354](https://github.com/xmidt-org/tr1d1um/issues/354)
@@ -210,7 +220,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial creation.
 
-[Unreleased]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.3...HEAD
+[Unreleased]: https://github.com/xmidt-org/tr1d1um/compare/v0.9.0...HEAD
+[v0.9.0]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.5...v0.9.0
+[v0.8.5]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.4...v0.8.5
+[v0.8.4]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.3...v0.8.4
 [v0.8.3]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.2...v0.8.3
 [v0.8.2]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.1...v0.8.2
 [v0.8.1]: https://github.com/xmidt-org/tr1d1um/compare/v0.8.0...v0.8.1